+++ src/index.ml
@@ -73,6 +73,8 @@ struct
     include Stats
   end
 
+  module Io_stats = Platform.Io_stats
+
   module IO = struct
     include Io.Extend (IO)