+++ src/stats.mli
@@ -24,7 +24,7 @@ type t = {
     - [time_sync] is the duration of the latest call to sync. *)
 
 val get : unit -> t
-val reset_stats : unit -> unit
+val reset : unit -> unit
 val incr_nb_merge : unit -> unit
 val incr_nb_replace : unit -> unit
 val incr_nb_sync : unit -> unit
@@ -35,3 +35,5 @@ module Make (_ : Platform.CLOCK) : sig
   val sync_with_timer : (unit -> unit) -> unit
   val add_merge_duration : Mtime.Span.t -> unit
 end
+
+module Io_stats (R : Platform.RAW_STATS) : Platform.IO_STATS with type t = R.t