+++ src/unix/raw_stats.mli
@@ -1,0 +1,4 @@
+include Index.Platform.RAW_STATS
+
+val add_read : t -> int -> unit
+val add_write : t -> int -> unit